requirements.txt
fastapi
uvicorn
numpy
Pillow
opencv-python
tqdm
matplotlib
torch==2.1.0
torchvision==0.16.0
--extra-index-url https://download.pytorch.org/whl/cu118
